requirements.txt
langchain==0.0.305
flask==2.3.3
openai==0.27.7
mysql-connector-python==8.1.0
pymysql==1.1.0